[Debug] moved special fatal error handlers to their own classes
